api-python/src/shared/openai/types/chat_completion_choice.py
from pydantic import BaseModel
from shared.openai.types.chat_message import ChatMessage


class ChatCompletionChoice(BaseModel):
    index: int
    message: ChatMessage
    finish_reason: str | None = "stop"
