api-python/pyproject.toml
[project]
name = "api-python"
version = "1.0.0"
dependencies = [
  "confluent-kafka==2.11.1",
  "fastapi==0.118.0",
  "httpx==0.28.1",
  "lancedb==0.25.1",
  "openai==1.109.1",
  "pydantic-ai-slim[openai]==1.0.10",
  "pydantic-settings==2.11.0",
  "sentence-transformers==5.1.1",
  "sentry-sdk[fastapi]==2.39.0",
  "uvicorn==0.37.0",
]

[dependency-groups]
dev = [
  "poethepoet==0.37.0",
  "pytest-cov==7.0.0",
  "pytest==8.4.2",
]

[tool.uv]
package = false
required-version = ">=0.8.0"

[tool.poe.tasks]
dev = { cmd = "python src/main.py", env = { ENV = "development" } }
serve = { cmd = "python src/main.py", env = { ENV = "production" } }
test = { cmd = "pytest --verbose --verbose", env = { ENV = "test" } }
test-coverage = { cmd = "pytest --cov=. --cov-report=xml", env = { ENV = "test" } }
